requirements.txt
matplotlib>=3.2.1
torchaudio
scipy
scikit-learn

numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
retina-face==0.0.14
seaborn==0.13.2

onnxruntime==1.19.0
uniface==0.1.7